refactor(plugin-rows): move hooks into hooks directory
